🧐 disable only iOS gesture, force fullscreen
only small fixes in this commit, will start working on the custom player next. currently, there is no way to go back to info on iOS; will be fixed in the custom player. also, autoPlay/autoInitialize options do not work for some reason - will investigate, so left them out of this commit.